Add some personal test images Consolidate classifyImage code to allow calling the marco classifier or the simple classifier by cmdline switch Move classifyImage training code into classifyImagesTrainer.py
