+++ src/bash/classifyImages.sh
@@ -6,8 +6,10 @@ function classifyImages {
     spark_hosts=($(cat ${PEGASUS_HOME}/tmp/crystal-project-spark-cluster/public_dns))
     postgresql_hosts=($(cat ${PEGASUS_HOME}/tmp/crystal-project-database-cluster/public_dns))
     ## copy the pyspark python script
-    echo "Copying python script to spark master..."
-    scp src/python/classifyImages.py ubuntu@${spark_hosts[0]}:
+    echo "Copying python scripts to spark master..."
+    for a in src/python/classifyImages{,{Marco,Simple}Partition}.py; do
+        scp "$a" ubuntu@${spark_hosts[0]}:
+    done
     ## launch the command
     echo "Running spark classification job..."
     ssh ubuntu@${spark_hosts[0]} "
@@ -15,6 +17,6 @@ function classifyImages {
         export LD_LIBRARY_PATH+=:/usr/local/hadoop/lib/native
         export SPARK_HOME=/usr/local/spark
         export HADOOP_HOME=/usr/local/hadoop
-        spark-submit --packages org.apache.hadoop:hadoop-aws:2.7.1,databricks:spark-deep-learning:0.3.0-spark2.2-s_2.11,org.postgresql:postgresql:42.1.4 --master spark://${spark_hosts[0]}:7077 --executor-memory 10G --driver-memory 40G --py-files classifyImages.py classifyImages.py $AWS_ACCESS_KEY_ID $AWS_SECRET_ACCESS_KEY $AWS_DEFAULT_REGION ${postgresql_hosts[0]} $POSTGRES_USER $POSTGRES_PASSWORD
+        spark-submit --packages org.apache.hadoop:hadoop-aws:2.7.1,databricks:spark-deep-learning:0.3.0-spark2.2-s_2.11,org.postgresql:postgresql:42.1.4 --master spark://${spark_hosts[0]}:7077 --executor-memory 30G --driver-memory 40G --py-files classifyImages.py,classifyImagesMarcoPartition.py,classifyImagesSimplePartition.py classifyImages.py $AWS_ACCESS_KEY_ID $AWS_SECRET_ACCESS_KEY $AWS_DEFAULT_REGION ${postgresql_hosts[0]} $POSTGRES_USER $POSTGRES_PASSWORD $1
         "
 }
